Created Slime enemy class and a dictionary of attacks + function to calculate attack damage
Continued abstraction of mechanics by making a global script containing the definition for Slime enemies and how their stats
scale with level, and creating a dictionary of attacks with definitions for how much damage they deal and of what type.
Implemented a basic function for calculating damage when a given attack is used by a character or enemy against a target.
This function currently only implements simple resistances functionality and treats every attack as working on the strength stat