Implemented initial components of battle system - player can use a single attack on a cooldown to attack a single opponent
Battle state loops according to a timstep (curently set to 0.02 seconds), allowing the player to act after a wind-up, and
set a new attack after a cool-down period, with the sum of these two times capped at 1 second.
There is currently no win condition.

Also added a GLobalUtilities script which for now contains a method for rounding to arbitrary values (e.g. nearest integer multiple of 0.02)